+++ src/main/java/com/example/raha/service/ArticleService.java
@@ -2,8 +2,6 @@ package com.example.raha.service;
 
 import java.util.List;
 
-import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
-import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import com.example.raha.domain.Article;
