+++ src/main/java/com/example/raha/service/SecurityUserService.java
@@ -1,11 +1,17 @@
 package com.example.raha.service;
 
+import java.time.OffsetDateTime;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpHeaders;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.algorithms.Algorithm;
 import com.example.raha.domain.User;
 import com.example.raha.repository.UserRepository;
 
@@ -13,13 +19,17 @@ import lombok.RequiredArgsConstructor;
 
 /**
  * SecurityConfigクラスで使用するユーザーサービス
+ * 
  * @author hosodatomoya
  */
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class LoginUserDetailsService implements UserDetailsService {
+public class SecurityUserService implements UserDetailsService {
     private final UserRepository repository;
+
+    @Value("${jwt.secret}")
+    private String secret;
 
     @Override
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
@@ -32,5 +42,20 @@ public class LoginUserDetailsService implements UserDetailsService {
                 .password(user.getPassword())
                 .build();
     }
-    
+
+    /**
+     * JWTヘッダーを生成
+     * 
+     * @param userId ユーザーID
+     * @return JWTトークン付きHttpヘッダー
+     */
+    public HttpHeaders createJwtHeader(Integer userId) {
+        // userIdクレーム、1日の有効期限を持つJWTトークンの生成
+        String token = JWT.create().withClaim("userId", userId)
+                .withExpiresAt(OffsetDateTime.now().plusDays(1).toInstant()).sign(Algorithm.HMAC256(secret));
+        HttpHeaders headers = new HttpHeaders();
+        headers.add("X-AUTH-TOKEN", "Bearer " + token);
+        return headers;
+    }
+
 }
